pyproject.toml
[tool.poetry]
name = "stepper-motor-driver"
version = "0.1.0"
description = "Package for the control of a 4-wire bipolar stepper motor"
authors = ["Armaan Rasheed <arasheed@uwaterloo.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sipyco = {git = "https://github.com/m-labs/sipyco.git", tag = "v1.8"}
numpy = "^2.1.3"
python-dotenv = "^1.0.1"
"RPi.GPIO" = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
